2354255 update the dio driver
git-svn-id: https://svn.code.sf.net/p/opensek/code/trunk@146 73ee5bea-b2a1-4b4e-8992-c3b1447581ab

+++ OpenDRV/Dio/inc/arm/lpc2468/Dio_Arch.h
@@ -83,7 +83,7 @@ typedef uint8 Dio_LevelType;
 
 /** \brief Port Level type definition */
 /* \req DIO024 Dio_PortLevelType is the type for the value of a DIO port. */
-typedef uint8 Do_PortLevelType;
+typedef uint32 Do_PortLevelType;
 
 /*==================[external data declaration]==============================*/
